# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Ranking API",
        "contact": {
            "name": "Nguyen Anh Duy",
            "email": "duynait@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1.0/ranking/get": {
            "get": {
                "tags": [
                    "Ranking"
                ],
                "summary": "Get ranking",
                "operationId": "3b4555c6c8482283f1fee6a9cfca64d2",
                "parameters": [
                    {
                        "name": "url",
                        "in": "query",
                        "description": "URL",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "keywords",
                        "in": "query",
                        "description": "Keywords",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ranking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid keywords. Maximum of 5 words, separated by new lines."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "URL not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "URL không tồn tại."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}